requirements.txt
flask
gunicorn
opencensus-ext-azure
applicationinsights